krep: bring the initial framework with help sub-command
The first two sub-commands "help" and "topic" are implemented
to output the summaries of the supported sub-commands and topics.

And extra optparse inherited from the lib to support "--no-"
prefix options, which suppress the option with an opposite or
negative meaning. krep is updated to handle the exception raised
by the inexistent option with a "--no-" prefix.

subcmd would work as the basic class of the sub-commands to be
implemented. logger provides a customized logging solution for
multiple threading with a local thread name. error would contain
the exceptions used by the sub-commands.