batch: add the sub-command for command batch handling
The command handles other commands in batch from the specific batch
files in the ini or XML format.

The sub-command with or without multiple thread will be executed
separately: The sub-command without thread will be executed in
parallel with the job number from the option, and then the sub-
commands with thread will be executed independently with the job
option, which could be handled by the command.

It has the features:

 - group filter to filter out the running projects
 - CLI options to build up the running command

Change-Id: If2a2e122dec9e1a532f6616706f87b6a194dd7de