config: add the component to read out config files
Add the class for config read, which supports two formats:

 - Extended ini format with sub-section support like git
 - XML configuration file with specified elements

The config file would keep the persist options for the tool.
It can be used as the config file of the tool or by the sub-
commands.

More information relative to the format can refer to the doc
of the file.

Change-Id: I61aee0925357309693c58a6fdbd37bd75e204563