file_utils: manage the package file extension in order
the file extension is important for the program to extract the
package file. for example, .tar.gz should be handled as a tgz
package instead of gzip package instead.

As the extensions in a hash table can't be kept well in order, it's
good to use a list instead to keep the extension in the purposed
order.

Change-Id: Idc7e44caff6e5f96f6db6801eb28ea7649756725