option: implement the formater for help with LF
In the design, the extra option will be built up with the strings
from each option providers in lists. And the LF will be used to
split each options but the default formatter will omit the LFs.

A new formater inherited from the default one is implemented to
handle the extra LF if the help string has the characters.

Change-Id: I8402be3e8571b5a7f8add8d4ba6f12924c2769ed