subcmd: support extra-option and extra_items
A class can work as a provider for extra items, which can be
defined like a two-value list:

 (
  (extra_option_1, extra_description_1),
  (extra_title_2, (
   (extra_option_2_1, extra_description_2_1),
   (extra_option_2_2, extra_description_2_2),
  )),
 )

extra_items from each provider will be formed and listed in
the option help under extra-option. As the extra description
wouldn't be formatted the overrided optparse formatter, it's
better not to use a long line for the description.

Change-Id: Ic9cd22a8eca6939cf5c053698047999d0205ebab