gerrit: use gerrit-cp:empty-commit to override default value
both argument initialize_commit and extra option
gerrit-cp:empty-commit can set up the command line to create
the empty commit when creating a new repository via gerrit.

to provide the possiblity that user can foribid the creation,
the extra option gerrit-cp:empty-commit supports to override
the argument - when its value is set to FALSE, the argument
will be ignore and the first empty commit wont' be create.

with the solution, a full clone of a repository with the default
"master" branch will be supported easily.

Change-Id: I5a33e44609f2756fb9a7d4c4d075e911cc92f770