krep: provide current-dir for the directory to run
the program will change to the working directory before
executing the sub-commands. with the assumption, the arguements
for files without absolute directory won't be right with the
working directory change.

new option current-dir will record the directory to run the
program, which can be used to join with the files from CLI.

Change-Id: Ia1550f8d550cfd050f51c840174f555579c5584c