pattern: distinguish continuable attribute for items
if a patten item contains more than one replacement, the continue
attribute will be described with the pattern one instead of the
last matched replacement item. the case excluded the situation
that continue attribute is added implicited.

with the limitation, the rule can be supported well to support
continue attribute within the replacements but terminated without
the pattern.

a sample is to use an xml pattern with comma-demilter rules, which
can be continuable or non-continuable. but the entire rule isn't
continuable at all.

Change-Id: I2d01deb11c9ec00d521fd1c19c72df104480d2e6