krep: handle colons in injection options
colon is the demiliter of patterns, and it's also the delimiter
of the prefix of inject options. it'll be hard to know the
meaning of the character if there's no extra judgement.

it's to use the characters like space and equaling sign as the
separator of the long options to make sure the colon is a part
of pattern or the option prefix.

Change-Id: If4b0361efb2f71f1a90f4be8e2e3e022647ebfe4