option: don't override the appending options
maybe it's a temperay solution not to override the option value,
which has the action "appending".

option "pattern" is the specific one to have the values from
command line, and xml files. the change fixes the case the pattern
options will override the values from xml files, which could
include lots of patterns, and are hard to handle alternatively
from the command line.

Change-Id: I0dd8c4febf1e595d306cb1fb70b98209dfb6cc40