git_project: allow to push SHA-1 as a head
git server secures to accept the valid branchname, but it's not
normal to use SHA-1 as the branch name.

A special case is to push the SHA-1 to a branch from a git-repo
static manifest. Now it's to support to SHA-1 to a "renamed"
branch with a replace-pattern when option "--force" is using.

 <rp-pattern category="rev">
   <pattern name=".+" replace="spec-name"/>
 </rp-pattern>

Change-Id: I407430308bda3ff0031a74064d873b02b647f2d0