config_file: support "value-sets" for key-value pairs
the xml element provides the possibility to enumerate the
key-value pairs in value-sets as source to parse xml nodes,
which can embed the referred variables.

with the solution, the config file can be split with the
configurable one and the one to provide the data. once the
data changed, only the one for data needs to be updated.
the maintenace effort will be more easier.

Change-Id: I19453b53579225e867e384ee38f75fa40e32c7d6