pattern: support category alias for pattern actions
as different categories contain different items, it's not possible
to reuse the items effectively with different category name.

a short name for category is good to use. and alias is the solution
to make category match several names including original name, and
its alias. every operation to the pattern will have a conversion
to validate the category with all aliases.

Change-Id: Iadf631a38610a58807a9bf2de52e9fd021204426